pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "rockstor"
version = "5.1.0"
description = "Btrfs Network Attached Storage (NAS) Appliance."
homepage = "https://rockstor.com/"
repository = "https://github.com/rockstor/rockstor-core"
documentation = "https://rockstor.com/docs/"
readme = "README.md"
keywords = ["linux", "NAS", "appliance", "btrfs"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Natural Language :: English",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3 :: Only", # at least initially
    # The "license" property should auto-invoke License classifier/classifiers.
]
authors = [
    "The Rockstor Project <support@rockstor.com>",
]
maintainers = [
    "Philip Paul Guyton <support@rockstor.com>",
    "Flox <support@rockstor.com>",
]
# Source0: (rockstor-core) is GPL-3.0-or-later, Source1: (rockstor-jslibs) has mixed licensing:
license = "GPL-3.0-or-later AND (MIT AND Apache-2.0 AND GPL-3.0-or-later AND LGPL-3.0-or-later AND ISC)"
include = [
    "COPYING",  # Our GPL 3 file from rockstor-core.
    "rockstor-jslibs.tar.gz",  # https://github.com/rockstor/rockstor-jslibs
    "rockstor-jslibs.tar.gz.sha256sum",  # sha256 of above tar.gz
    "build.sh",  # master build script
    "poetry.toml",  # poetry config
    "poetry.lock",  # current poetry established dependency lock file.
    ".env",  # poetry-plugin-dotenv default source file.
    { path = "conf" },  # Configuration directories
    { path = "etc" },
    { path = "var" },  # Some processes depend on this tree existing.
]
packages = [
    { include = "rockstor", from = "src" },
    { include = "storageadmin", from = "src/rockstor" },
    { include = "smart_manager", from = "src/rockstor" },
]

[tool.poetry.build]
generate-setup-file = false
# Future capability to run a python script first when 'poetry build' is called.
# https://github.com/python-poetry/poetry/issues/2740
# script = "build.py"

[tool.poetry.dependencies]
# We use groups (>=1.2.0) to clarify our dependencies (essentially labels):
# We use to need Poetry 1.1.15 for it's Python 2.7/3.6 compatibility.
# https://python-poetry.org/history/#120b1---2022-03-17
# 1.20b1 dropped Python 3.6 support.
# https://python-poetry.org/docs/managing-dependencies/#dependency-groups
#
# https://python-poetry.org/docs/dependency-specification
python = "~3.11"

# [tool.poetry.group.django.dependencies]
django = "~4.2"
django-oauth-toolkit = "~2.4"
# DRF next version (3.16.0) changes behaviour!
djangorestframework = "==3.15.2"
django-pipeline = "==4.0.0"
docutils = "~0.21"
python-engineio = "==4.8.0"
python-socketio = "==5.9.0"
# Need dbus-python 1.4.0 for Py3.13
dbus-python = "==1.3.2"
psycopg = "~3"
psycogreen = "==1.0"
# gevent 24.10.[1 & 2] resulted in: AUTH_REQ_SASL_CONT without AUTH_REQ_SASL
gevent = "24.2.1"  # can be an extra dependency to gunicorn.
gunicorn = "~23"
packaging = "*"  # 'btrfs version' parsing.

# [tool.poetry.group.tools.dependencies]
six = "==1.16.0"  # 1.14.0 (15 Jan 2020) Python 2/3 compat lib
huey = "*"
psutil = "==5.9.4"
# mock = "==1.0.1" now part of std lib in Python 3.3 onwards as unittest.mock
# pyzmq requires libzmq5 on system unless in wheel form.
pyzmq = "*"
distro = "*"
URLObject = "==2.1.1"
keyring-pass = "*"
zypper-changelog-lib = "0.7.9"
# zypper-changelog-lib = {path = "/path/zypper-changelog-lib/dist/zypper_changelog_lib-0.7.9-py3-none-any.whl"}
# https://pypi.org/project/supervisor/ 4.1.0 onwards embeds unmaintained meld3
supervisor = "==4.2.4"

# `poetry install --with dev`
[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "*"

[tool.poetry.scripts]
# https://python-poetry.org/docs/pyproject#scripts
# script-name = package.module.function
# -- Init and startup related --
initrock = 'scripts.initrock:main'  # systemd rockstor-pre
flash-optimize = 'scripts.flash_optimize:main'
prep_db = 'scripts.prep_db:main'
bootstrap = 'scripts.bootstrap:main'  # systemd rockstor-bootstrap
qgroup-clean = 'scripts.qgroup_clean:main'
qgroup-maxout-limit = 'scripts.qgroup_maxout_limit:main'
# -- Web-UI --
data-collector = 'smart_manager.data_collector:main'
replicad = 'smart_manager.replication.listener_broker:main'
send-replica = 'scripts.scheduled_tasks.send_replica:main'
backup-config = 'scripts.config_backup:main'
mnt-share = 'scripts.mount_share:mount_share'  # Samba mnt_helper
st-pool-scrub = 'scripts.scheduled_tasks.pool_scrub:main'
st-snapshot = 'scripts.scheduled_tasks.snapshot:main'
st-system-power = 'scripts.scheduled_tasks.reboot_shutdown:main'
delete-api-key = 'scripts.delete_api_key:main'
pwreset = 'scripts.pwreset:main'
ovpn-client-gen = 'scripts.ovpn_util:client_gen'
ovpn-client-print = 'scripts.ovpn_util:client_retrieve'
ovpn-initpki = 'scripts.ovpn_util:initpki'
# -- CLI Utilities --
debug-mode = 'scripts.debugmode:main'
delete-rockon = 'scripts.rockon_delete:delete_rockon'
# qgroup-test = 'scripts.qgroup_test:main'  # broken, in need of update/repair.

# Legacy scripts
# rcli = 'cli.rock_cli:main'
# docker-wrapper = 'scripts.docker_wrapper:main'
# rockon-json = 'scripts.rockon_util:main'

